Refactored test and added gr test
