+++ tests/.gitignore
@@ -1,3 +1,3 @@
 !.gitignore
-!test_*.py
+!swiftest_tests.py
 simdata/**
